requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
Pillow>=10.0
requests>=2.31
python-telegram-bot>=21.0
gunicorn>=22.0
